Fix float → int panic on /admin
As row count (reltuples) got high (> 1 mil) and because the column is a
float, postgres started to return the value in e notation which Go
panicked when trying to write to an int. Fix with a cast.

+++ routes/admin.go
@@ -44,9 +44,9 @@ func adminGET(w http.ResponseWriter, r *http.Request) {
 
 	if err := db.Select(
 		&data.Tables,
-		` SELECT relname                                           name,
-		         CASE WHEN relkind = 'i' THEN 0 ELSE reltuples END rows,
-		         PG_TOTAL_RELATION_SIZE(c.oid)                     size
+		` SELECT relname                                                name,
+		         CASE WHEN relkind = 'i' THEN 0 ELSE reltuples::int END rows,
+		         PG_TOTAL_RELATION_SIZE(c.oid)                          size
 		    FROM pg_class     c
 		    JOIN pg_namespace n
 		      ON n.oid = relnamespace
